refactor(docs-framework): fold lustro-search in, retire it as a package
Moves @eventuras/lustro-search's source into docs-framework/src/search (git
renames preserve history) and exposes it via new ./search, ./build-index and
./react subpaths. docs-framework now builds with vite (react-lib preset) and
declares @orama/* deps plus @eventuras/ratio-ui and react as optional peers.

dev-docs consumes the new subpaths; the lustro-search package and its repo.json
entry are removed. Adopting the standard react-lib tsconfig surfaced three
strict-null issues in collect.ts, fixed with no behaviour change.
