+++ libs/docs-framework/src/search/build-index.ts
