+++ libs/docs-framework/src/search/index.ts
