+++ libs/docs-framework/src/search/react.ts
