+++ libs/docs-framework/src/search/schema.ts
