+++ libs/docs-framework/src/search/types.ts
