+++ libs/docs-framework/vite.config.ts
@@ -3,12 +3,14 @@ import { defineReactLibConfig } from '@eventuras/vite-config/react-lib';
 export default defineReactLibConfig({
   entry: {
     index: 'src/index.ts',
-    react: 'src/react.ts',
-    'build-index': 'src/build-index.ts',
+    'search/index': 'src/search/index.ts',
+    'search/build-index': 'src/search/build-index.ts',
+    'search/react': 'src/search/react.ts',
   },
   external: [
     '@orama/orama',
     '@orama/plugin-data-persistence',
+    'fast-glob',
     /^@eventuras\//,
     /^node:/,
   ],
