--- libs/lustro-search/CHANGELOG.md
@@ -1,70 +1,0 @@
-# @eventuras/lustro-search
-
-## 4.0.4
-
-### Patch Changes
-
-- Updated dependencies [3543c98]
-  - @eventuras/ratio-ui@1.0.4
-
-## 4.0.3
-
-### Patch Changes
-
-- 7c9fe79: chore: update dependencies
-- Updated dependencies [7c9fe79]
-  - @eventuras/ratio-ui@1.0.3
-
-## 4.0.2
-
-### Patch Changes
-
-- Updated dependencies [e0b00a9]
-  - @eventuras/ratio-ui@1.0.2
-
-## 4.0.1
-
-### Patch Changes
-
-- Updated dependencies [e073558]
-  - @eventuras/ratio-ui@1.0.1
-
-## 4.0.0
-
-### Patch Changes
-
-- Updated dependencies [abaa171]
-- Updated dependencies [202f819]
-- Updated dependencies [7b0c54c]
-  - @eventuras/ratio-ui@1.0.0
-
-## 3.0.1
-
-### Patch Changes
-
-- Updated dependencies [d5634da]
-  - @eventuras/ratio-ui@0.14.1
-
-## 3.0.0
-
-### Patch Changes
-
-- Updated dependencies [bbb9111]
-- Updated dependencies [0e1796e]
-  - @eventuras/ratio-ui@0.14.0
-
-## 2.0.0
-
-### Patch Changes
-
-- Updated dependencies [0b4b869]
-  - @eventuras/ratio-ui@0.13.0
-
-## 1.0.0
-
-### Patch Changes
-
-- Updated dependencies [fce9a48]
-- Updated dependencies [cc205db]
-- Updated dependencies [21d0d6f]
-  - @eventuras/ratio-ui@0.12.0
