--- libs/lustro-search/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@eventuras/lustro-search",
-  "version": "4.0.4",
-  "private": true,
-  "description": "Full-text search abstraction powered by Orama — shared across docs, web, and other apps",
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    },
-    "./react": {
-      "types": "./dist/react.d.ts",
-      "import": "./dist/react.js"
-    },
-    "./build-index": {
-      "types": "./dist/build-index.d.ts",
-      "import": "./dist/build-index.js"
-    }
-  },
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "vite build",
-    "dev": "vite build --watch"
-  },
-  "dependencies": {
-    "@orama/orama": "^3.1.18",
-    "@orama/plugin-data-persistence": "^3.1.18"
-  },
-  "devDependencies": {
-    "@eventuras/typescript-config": "workspace:*",
-    "@eventuras/vite-config": "workspace:*",
-    "@types/node": "^24.12.0",
-    "@types/react": "^19.2.14",
-    "react": "^19.2.5",
-    "vite": "^8.1.0"
-  },
-  "peerDependencies": {
-    "@eventuras/ratio-ui": ">=1",
-    "react": "^18 || ^19"
-  },
-  "peerDependenciesMeta": {
-    "@eventuras/ratio-ui": {
-      "optional": true
-    },
-    "react": {
-      "optional": true
-    }
-  }
-}
