--- libs/lustro-search/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@eventuras/typescript-config/react-library.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "types": ["react", "react-dom", "node"]
-  }
-}
